requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.4.1
Django==3.2.9
django-cors-headers==3.10.0
django-filer==2.1.2
django-filter==21.1
django-js-asset==1.2.2
django-mptt==0.13.4
django-polymorphic==3.0.0
djangorestframework==3.12.4
drf-yasg==1.20.0
easy-thumbnails==2.8
Faker==10.0.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.3
importlib-metadata==4.8.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
lxml==4.6.4
Markdown==3.3.5
MarkupSafe==2.0.1
mixer==7.2.0
packaging==21.3
Pillow==8.4.0
promise==2.3
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
reportlab==3.6.2
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
svglib==1.1.0
text-unidecode==1.3
tinycss2==1.1.0
Unidecode==1.1.2
uritemplate==4.1.1
urllib3==1.26.7
webencodings==0.5.1
zipp==3.6.0
